requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.4
jinja2==3.1.4
pyyaml==6.0.2
loguru==0.7.2
scikit-learn==1.5.1
rank-bm25==0.2.2
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
openai==1.51.2

